refractored StartMenue into own class
